COMP: Fix "signed/unsigned" warning in qSlicerVolumeRenderingSettingsPanel
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=20692
git-svn-id: http://svn.slicer.org/Slicer4/trunk@20692 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/Loadable/VolumeRendering/qSlicerVolumeRenderingSettingsPanel.cxx
@@ -194,7 +194,7 @@ void qSlicerVolumeRenderingSettingsPanel
   const std::map<std::string, std::string>& renderingMethods =
     d->VolumeRenderingLogic->GetRenderingMethods();
   /// \todo not the best test to make sure the list is different
-  if (renderingMethods.size() != d->RenderingMethodComboBox->count())
+  if (static_cast<int>(renderingMethods.size()) != d->RenderingMethodComboBox->count())
     {
     std::map<std::string, std::string>::const_iterator it;
     for (it = renderingMethods.begin(); it != renderingMethods.end(); ++it)
